serialize works for client server

+++ pdu_handler/pdu_helper.c
@@ -86,7 +86,7 @@ int open_fd(char *path){
 }
 
 void read_from_fd(int fd, void *destination, int size) {
-    if(read(fd, destination, size) < 0){
+    if(read(fd, destination, (size_t) size) < 0){
         perror_exit("read()");
     }
 }
